protocol: grpc: Allow single storage entry for CreateSandbox()
We don't need a storage list in case of CreateSandbox(). This function
is a very particular case where we are passing a shared directory via
9pfs. This shared directory is the way to share some specific files
and directories between host and guest. This is different from
CreateContainer() function where we expect several devices and
volumes to be passed to the VM.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>