protocol: grpc: Add Flags parameter to Storage structure
Basically, the storage structure is being used to describe a mount
that should be performed prior to the container launching. And if
we take a look at the parameters needed to perform a mount, flags
is one of them.

The idea is that our agent will expose a map of string in order to
describe the supported flags. This way, it will translate those
strings into actual flags understood by the guest OS.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>