protocol: grpc: Allow several updates of the same interface
When using UpdateInterface(), the way the UpdateType is defined only
allows for one operation at a time. This commit modifies the enum
values of the UpdateType so that they can be used as flags. This way,
we can combine several updates (IPs, Name, MTU) of the same interface
in one call.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>