agent: Add shared PID namespace support
We want to be able to share the same PID namespace across all
containers of a sandbox. This commit allows for such a thing by
starting a new process from a pause binary, into a new PID ns.

This way, the process is considered as init process inside this
namespace and the namespace will exist as long as this process
runs.

Fixes #20

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>