mount: Handle storage options as flags and options combination
The Options field inside Storage structure is used to pass both
flags and data to the mount() system call. The semantic behind
this actually checks if an option from the list matches with a
flag from a specific list of flags. In case it does not match
any flag, the option is considered as a real option that should
be used as data parameter.

Fixes #26

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>