network: Allow several updates on the same interface
With the recent protocol change for the list of possible actions
updating a network interface, it is now possible to combine several
actions in only one call to UpdateInterface(), by adding those actions
as flags.

This commit implements the support for this feature. It returns an
error in case none of the known actions are parsed from the action
type. Indeed, we should not call into UpdateInterface() if there is
nothing to update.

Fixes #42

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>