make UpdateInterface() declarative
And add interface_id to identify interfaces.

UpdateInterface() is using bitwise operation which is
too lowel and it doesn't support remove&add IPs at the
same grpc request. So it is changed to declarative way.

On declarative way, the agent should roll back when failed.
Rolling back can also fail again, the runtime need to know
the final result of the interface in this case, so the API
also returns the interface regardless its success or not.

Signed-off-by: Lai Jiangshan <jiangshanlai@gmail.com>