Add AddDevices()/RemoveDevices() API.
The type Device is converted from Storage.
And a device_id is added to indentify the devices.
The missing `driver_options` is also added.

OnlineCPUMem() is removed since we can use the
AddDevices()/RemoveDevices() for it.

Also move the storage operation out from CreateContainer(),
since we can also use the AddDevices()/RemoveDevices() for it.
It makes CreateContainer() more like oci `runtime create`.

And add some comments in the .proto.

Signed-off-by: Lai Jiangshan <jiangshanlai@gmail.com>