Always create sandbox init process and remove the field sandbox_pidns
In hyperstart, the pod_init process has its own net/pid/uts/ipc ns.

And in kata, these namespaces are also essential.

Net ns for the sandbox should be always created since the container's
network should be isolated from the vm's initial network to
avoid messing.

Pid ns for the sandbox should be always created since the containers
always join this ns. We can use new pid ns for each container,
but in most cases, they are shared the same sandbox init pid ns.

Uts ns for the sandbox should be always created since the hostname
is set in it without mess it up with the vm's initial UTS.

Ipc ns for the sandbox should be always created since the containers
always join this ns. We can use new ipc ns for each container,
but in most cases, they are shared the same sandbox init ipc ns.

So sandbox init process is better to be always created
in its own net/pid/uts/ipc ns and let the later containers
join in. To make it different from hyperstart, we can use
some config to tell the container not to join it in
some cases later.

Signed-off-by: Lai Jiangshan <jiangshanlai@gmail.com>