protocols: grpc: Rely on exec_id provided by the runtime
This commit radically changes the way the process IDs are generated.
Indeed, for simplicity, the process ID, called exec_id, will always
be provided by the runtime.

The shim needs to be started knowing which exec_id it actually
represents. By having this information from the runtime, we make
things simple since the runtime will provide this directly to the
shim, with no need to interact with the agent.

Also, this commit stores the exitCode channel from the caller, so
that we don't need to keep it into a specific map for too long. This
prevents from the issue of continuing to use a map with a specific
PID entry, while this PID could have been re-assigned by the kernel.

Fixes #82

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>