grpc: Move sequencing from StartContainer to CreateContainer
With recent changes and in order to support the shim connection
after the container has been created, we need to initialize the
exit code channel and to setup the terminal from the container
creation step. StartContainer becomes a very simple function and
actually fits completely the expected libcontainer behavior.

Fixes #102

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>