agent: Fix the amount of bytes transmitted to the shim
Every time our GRPC functions ReadStdout() or ReadStderr() were
called, the entire allocated buffer was returned. This is wrong
since we don't know the amount of bytes we will get on both stdout
and stderr. Instead, this patch returns the slice corresponding to
the real amount of bytes read from the process.

Fixes #123

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>