grpc: Handle hotplugged devices
This commit adds a new Device structure in order to avoid confusion
between what should be mounted and what is a device. When creating
a new container, we expect the caller to pass a list of storages,
and a list of devices.

When adding a host device to the container by using things like
"docker run --device /dev/sda1:/dev/blk1", the runtime will not
be able to anticipate the major/minor numbers for the device
being hotplugged into the VM. This means the agent has to update
the list of devices provided through the spec, by relying on the
list of devices provided.

Fixes #132

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>