grpc: network: add UpdateRoutes
To provide a true declarative interface, it is important that we
introduce an UpdateRoutes function.

This function will clear out all existing routes and then attempt to
apply what has been passed over gRPC. In the event that a failure is
observed when applying routes, the error is logged, message is returned
as well as the final state of the ip routes in the guest.

The routes structure in the gRPC prtocol was enhanced in order to
provide handling of Source and Scope fields, which are necessary to be
able to apply functional fresh routes.

Fixes: #149

Signed-off-by: Eric Ernst <eric.ernst@intel.com>