agent: Allow to retrieve the process of a stopped container
A container being in stopped state is not a valid reason for not
accepting a call to getProcess(). Indeed, the explanation behind
this is that we want to make sure a shim can connect after the
process has already terminated. Some processes have a very short
lifetime and the shim might end up calling into WaitProcess()
after that. This does not mean that we cannot retrieve the output
and the exit code, they are waiting for the shim for that. But we
need to let the shim by not returning an error in this case.

Fixes #145

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>