storage: Rework the way Storage and Device are handled
Storage and Device structures were not used the way they should
have been used, that is match with OCI Mounts and Devices.
To give more detail about this, OCI considers as Devices what is
passed through options such as "--device" and Mounts what is passed
through options as "--volume".
For each volume, an entry is going to be created into the OCI spec,
and we should create the appropriate Storage structure accordingly.
For each device, an entry is going to be created in the OCI spec,
and we should create the appropriate Device structure accordingly.
Now, rootfs could be a specific case, but instead, we consider it
as a Storage entry, meaning it will be mounted without any difference
compared to regular volumes.
Eventually, the libcontainer is called after all mounts have been
performed, bind mounting rootfs first, based on oci.Root.Path and
then the volumes listed as oci.Mounts and the devices listed as
oci.Device.

Fixes #166

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>