device: Add SCSI block device support
Recently, the support for block devices (rootfs or support for
devices) plugged through the VM with virtio-blk have been added.

This commit follows this by allowing block devices plugged through the
VM using virtio-scsi to be supported and properly processed from the
agent.

One thing to notice here is that latest versions of Qemu (>= 2.9) have
some stability issues using virtio-blk, and the use of virtio-scsi is
a good way to work around this current limitation.

Fixes #162

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>