tests: Skip more tests if non-root
The unit tests in this repository are ultimately run by the
`go-test.sh` script in the `tests` repository. This script actually runs
the unit-tests *twice* **if** it is invoked as a non-`root` user: once
as the non-`root` user and once as `root`. This is done to cover all
possible scenarios: certain tests require super-user privileges whilst
other tests are only valid when running as a non-privileged user. The
script then combines the test results to calculate a coverage
percentage.

The Travis test runs are currently failing (in the first non-`root`
"run") as some of the tests require root privileges.

Make extra calls to `skipUnlessRoot()` for those failing tests to allow
them to be skipped when run as a non-`root` user and then run as `root`
in the second "run".

Fixes #177.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>