grpc: update cpuset cgroup
Each time a container is created, it is placed inside a cpuset cgroup
with the actual number of vCPUs, when more vCPUs are hot added this cgroup
MUST BE updated to allow containers accessing to these new resources. The
cpuset cgroup is not updated if the container already have one defined by user.

fixes #219

Signed-off-by: Julio Montes <julio.montes@intel.com>