block: Use PCI address to determine block device name
Use the PCI address identifier passed in the Id field of
Device struct to get the device name of the block device
passed with blk driver. The runtime does not have to predict the
device name any longer.

Use the PCI address to check if hotplug event has been received
by checking in the device map. If the PCI address is not present
in the map, add a watcher to the list of watchers to get
notification when the uvent is received.

We need to lock the sandbox while checking in the sandbox
map, so pass sandbox as an additional parameter to device
handlers.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>