channel: support communication channel hotplug
Since communication channel (vsock or serial port) is hot plugged by the
runtime, newChannel iterates in a loop looking for the serial port or vsock
device. The timeout is defined by channelExistMaxTries and channelExistWaitTime
and it can be calculated by using the following operation:

```
(channelExistMaxTries * channelExistWaitTime) / 1000 = timeout in seconds
```

If there are neither vsocks nor serial ports, an error is returned.

fixes #298

Signed-off-by: Julio Montes <julio.montes@intel.com>