agent: write OCI spec to config.json in bundle path
In order to comply with the OCI specification, the
spec must be written to a file named 'config.json'
located in the bundle directory. To properly set the
bundle path at the time of libcontainer config creation
the cwd also needs to be changed to the bundle path.
CreateContainer was refactored to reduce cyclomatic
complexity.

Fixes: #349

Signed-off-by: Edward Guzman <eguzman@nvidia.com>