uevent:Revert "device: check for existing device PCI path before waiting"
We should not check for the sysfs path assciated with the PCI address
as the device node may still not be created. Only a uevent with a device
name associated indicates that the device is ready.

This change was also causing the sandbox lock to be not released causing
the uevent listening loop to be locked permanently.

This reverts commit 7caf1c8bd9ebad9ab6631e1839d62c8c98ed7d8a.

Fixes #379

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>