agent: add memory hotplug probe interface check in GetGuestDetails
There are 2 phases in Memory Hotplug. 1) Physical Memory Hotplug phase.
2) Logical Memory Hotplug phase.
The First phase is to communicate hardware/firmware and make/erase
environment for hotplugged memory.
If firmware supports notification of connection of new memory to OS,
this phase is triggered automatically. ACPI can notify this event.
This also what kata supports on amd64, memory hotplug via acpi-driven.
But if not, there is another option, probe operation by hand.
And since memory hotplug via acpi is missing on qemu-system-aarch64,
we hope to support the other probe solution.
This whole implementation of memory hotplug via probe interface is
divided into two phases, the first is to check if guest kernel
is capable of memory hot-add via probe interface, which is located at
/sys/devices/system/memory/probe.

Fixes: #442

Signed-off-by: Penny Zheng <penny.zheng@arm.com>