client: Dont try connect to non existing socket.
When a new Client is created, it tries to connect to
a socket no matter if exist or not. This is not
and issue because the grpc api will time out after
some time. But some simple task like kata-runtime list can
take a lot of the socket does not exist because the
VM and proxy top working.

Fixes: #453

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>