wip: network: Wait till link is operationally up.
Although we set the link up with LinkSetup, this function
returns before the link is actually up.
Additionally, besides the admin state of a network link, we
also need to check if the link is operationally up.
Namely we need to check if the netlink attribute IFLA_OPERSTATE is set
to IF_OPER_UP.

Fixes #458

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>