tracing: Hide trace messages unless tracing enabled
Don't log trace-related messages unless tracing is actually enabled. If
tracing is not enabled, trace spans are still created, but the tracer is
a NOP. However, this is an implementation detail the user does not need
to be aware of so hide trace logs unless the user has enabled tracing
explicitly.

Fixes #482.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>