agent: Add support for local storage
Local storage is just a directory created inside the VM.
This will use whatever the storage driver is for the container
rootfs. This will normally be 9p, but in some cases could be
device mapper. In this case the directory will benefit from the improved
performance of device mapper.

Fixes #524

Signed-off-by: Alex Price <aprice@atlassian.com>