agent: support debug console
Start a debug console if the `agent.debug_console` option is part of the kernel
command line and a shell (bash or sh) is installed in the rootfs.
The debug console is useful when the developer wants to debug the agent process
or any other processes (even containers). The current way to do this is by
modifying a systemd target + creating a new system unit, the problem with this
approach is that it's not user-friendly and we depend on systemd, since it must
be installed in the rootfs.

fixes #546

Signed-off-by: Julio Montes <julio.montes@intel.com>