console: Add debug console test
Changed `setupDebugConsole()` to use a global `var` of supported shells
and to accept the debug console path as a parameter. This allows the
new unit test to modify its behaviour.

Note that to allow the goroutine started by `setupDebugConsole()` to be
stopped, the function now requires a `context` parameter. That required
`realMain()` to call the debug console setup after starting tracing (to
allow the tracing context to be passed to it). This also has the
attractive side-effect of allowing the debug console setup itself to be
traced.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>
(cherry picked from commit 5163babcedd1d7d5c29e62266838773861cab5b8)
Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>