device: Allow uevent handler to be stopped
Pass a context to `waitForDevice()` to allow the goroutine started by
this function to be stopped. This specifically for testing
purposes.

Due to how and where `waitForDevice()` is called, this change required
adding a `context` parameter to a number of other functions, notable to
the `deviceHandler` and `storageHandler` functions.

Also moved the `timeoutHotplug` `const` variable to `device.go` (where
it is actually used) and made it a true `var` to allow the tests to
manipulate it.

Enabling the stopping of the uevent handler also allowed a lot of new
unit tests to be added.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>