device: fix the issue of failed waiting on device appeared in /dev
There is an issue with the waitForDevice:
this wait will first check is the device's sys entry existed in /sys directory,
and if it existed, then it would return directly. How ever, the kernel register
a new device as an order of: 1) create the sys entry in /sys; 2) mknod in devtmpfs;
3) send uevent to userspace. The problem here is when the "wait" tested the sys entry
existed and return directly, but at that time the kernel would haven't created the device
node in devtmpfs, thus the following mount failed with couldn't resolve the device path.

Thus, the correct choice for waiting the device is depends on the uevent notice, since
once the kernel sended the device "add" uevent, it had created the device node in devtmpfs
yet.

Fixes: #628

Signed-off-by: lifupan <lifupan@gmail.com>
(cherry picked from commit 006fdfecbb2b69132c0156714612688c11f9c2cf)
Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>