mount: ensure local directory storage types have the correct permissions
This commit ensures that local directory storage types have the correct
permissions set on the mount point. We can't rely on os.MkdirAll to
set the desired permissions, as the agent process umask will determine
the permissions on the created directory. To fix this we call os.Chmod
on the mount point with the desired permissions.

Additional tests have been added.

Fixes #636

Signed-off-by: Alex Price <aprice@atlassian.com>