agent: add support for logging to a vsock port
not all the hypervisors connect a socket to /dev/console, so it's no possible
to get the agent's logs from a unix socket. This patch is to connect the
logger (logrus) to a vsock port, this way the logs can be read in the
host using vsocks or hybrid vsocks.

fixes #652

Signed-off-by: Julio Montes <julio.montes@intel.com>