client.go: HybridVSockDialer: Change Read EOT to recv peek
Rust agent(grpc-rs) cannot work with firecacker because the first package
will be dropped or truncated by HybridVSockDialer read EOT code.
This issue was reported in
https://github.com/kata-containers/kata-containers/issues/79

Golang agent(grpc-golang) is also affected by this issue.  But it is first
package is 9 bytes FRAME_SETTINGS.  Drop it will not affect subsequent
communications.

This commit change to use recv peek that will do same check with read
EOT but not drop any package.

Fixes: #695

Signed-off-by: Hui Zhu <teawater@antfin.com>