client.go: HybridVSockDialer: Check return size n of unix.Recvfrom
I found that unix.Recvfrom will not return err to io.EOF if it get
an EOF.  It just set return size n to 0.
That will make function commonDialer return. Grpc dialer will handle
it.

This commit add code to check return size n and set err to io.EOF if
n is 0.

Fixes: #697

Signed-off-by: Hui Zhu <teawater@antfin.com>