protocols: client: Increase connection retry timeout
The client needs more than 15s to retry connecting to the agent since
we know the agent might take some time to be ready and listen onto
vsock. Indeed, in the context of the CI running nested virtualization,
the VM might take some time to be spawned and it might boot a bit slower
than expected, which means we need to let more time to the agent to
listen on the AF_VSOCK socket.

Bumping to 60s is acceptable since it will still make the container
fails fairly quickly if things are really broken for some other reasons.

Fixes #732

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>