protocols: client: Re-enable proper retry strategy
Now that sporadic errors have been fixed regarding the combination of
Kata Containers with Cloud-Hypervisor, we can go back to the proper
handling of the connection error. In case the connection fails, we don't
want to rely on the slow backoff strategy from GRPC (takes 1s before it
retries). Instead, we want to rely on our custom handling of retrying to
connect as soon as possible.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>