device: Accept the PCIAddress in the BDF format for block devices
The kata-agent now only accepts the PCIAddress from kata-runtime in
the "bridgeAddr/deviceAddr" format, which is tied to the device
hotplug interface in qemu. This patch extends the kata-agent to accept
input PCIAddress in the standard BDF format, which should be more
generic to support different hypervisors, e.g. cloud-hypervisor.

Fixes: #823

Depends-on: github.com/kata-containers/runtime#2909

Signed-off-by: Bo Chen <chen.bo@intel.com>