device: Revert: Allow to use the predicted 'VmPath' when adding blk devices
This reverts commit c01192e76a139c9bd18d6bedef9ed9554dc5cd1e.
The above commit was a temporary (not reliable) workaround to support
container rootfs through hotplugged block device when using
cloud-hypervisor (clh). As the updated version of clh now returns the
PCI BDF information for all hotplugged devices, we no longer need to
rely on the predicted device path in the guest (a.k.a 'VmPath'). This
patch reverts the above commit to remove the special code path for
supporting cloud-hypervisor.

Fixes: #823

Signed-off-by: Bo Chen <chen.bo@intel.com>