device: Explicitly wait for VFIO devices to be ready
If extra devices are given in the OCI spec, the user will expect them
to be ready and available once the (inner) container starts executing.
So, if that means hotplugging something into the Kata VM, we need to
wait for that hotplug operation to complete before executing the
container.  For most devices that's handled by getDeviceName() which
waits for a uevent indicating the device is ready.

VFIO devices, however, don't have any handler so we don't explicitly
wait for them to be ready.  We usually get away with it, because we
force a PCI rescan in in finishCreateContainer() which blocks until
the rescan is complete.  That really only works by accident though,
completing a rescan doesn't necessarily mean all the device probing
logic is complete.  Worse, in some cases the PCI rescan can collide
with the hotplug processing and cause the device to go into a broken
state.

So, instead of relying on the rescan, accept information from the
runtime about what VFIO devices we expect, and explicitly wait for
them to be ready.

VFIO devices in Kata are (for now) weird - they will bind to whatever
the Kata VM's driver for them is, so could appear as any kind of
device (char, block, network interface, etc.).  That means waiting for
just the PCI device to be ready isn't foolproof, since there could be
some extra processing time for the driver to complete probing and
create the appropriate secondary devices.

But, really, that was true of the forced rescan as well, so this is a
clear improvement.

fixes #781
Depends-on: github.com/kata-containers/runtime#2981

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>