device: Refactor updateSpecDeviceList
updateSpecDeviceList is used to update the container configuration to
change devices major/minor numbers configured by the Kata client based
on host details to values suitable for the sandbox VM, which may
differ.  It takes a 'device' object, but the only things it actually
uses from there are ContainerPath and VmPath.

Refactor this as updateSpecDevice(), taking the host and guest paths
to the device as explicit parameters.  This makes the function more
self-contained and will enable some future extensions.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>