device: Index all devices in spec before updating them
The agent needs to update device entries in the OCI spec so that it has the
correct major:minor numbers for the guest, which may differ from the host.

Entries in the main device list are looked up by device path, but entries
in the device resources list are looked up by (host) major:minor.  This is
done one device at a time, updating as we go in updateSpecDeviceList().

But since the host and guest have different namespaces, one device might
have the same major:minor as a different device on the host.  In that case
we could update one resource entry to the correct guest values, then
mistakenly update it again because it now matches a different host device.

To avoid this, rather than looking up and updating one by one, we make all
the lookups in advance, creating a map from (host) device path to the
indices in the spec where the device and resource entries can be found.

Fixes: #834

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>