device: Rename pciDeviceMap in sandbox struct
pciDeviceMap is essentially just a map from sysfs paths to corresponding
device nodes.  The only thing PCI related is that the sysfs paths we're
usually looking up in it usually involve PCI addresses.  There's no reason
that has to be the case, however, and I have future uses which will not
be PCI.

Therefore, rename it (and change some surrounding comments) to reflect
its more general nature.  While we're there, fix a typo in a related
comment.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>