device: Increase hotplug timeout
hotplugTimeout controls how long we'll wait for a uevent indicating that a
device is ready in getDeviceName().  It's currently 3s, which is plenty of
time to allow for udev processing that we're usually dealing with.

However, we have upcoming cases where we may need to wait for a complete
SHPC PCI hotplug to complete.  SHPC has a 5s delay built into the protocol
(and therefore the guest implementations).

To accomodate this case, increase the timeout to 10s.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>