mount: don't error of virtiofs share is already mounted
In the event that the virtiofs device is already mounted at the
requested destination, don't error out. We'll check before attempting to
mount to see if the destination is already a mount point. If so, skip
doing the mount in the agent.

This facilitates mounting the sharedfs automatically in the guest before
the agent service starts.

Fixes: #884

Signed-off-by: Eric Ernst <eric.g.ernst@gmail.com>